+++ src/main/MQ2Commands.cpp
@@ -4248,7 +4248,7 @@ void AltAbility(SPAWNINFO* pChar, char* szLine)
 
 			for (int nAbility = 0; nAbility < AA_CHAR_MAX_REAL; nAbility++)
 			{
-				if (ALTABILITY* pAbility = GetAAByIdWrapper(pLocalPC->GetAlternateAbilityId(nAbility)))
+				if (CAltAbilityData* pAbility = GetAAById(pLocalPC->GetAlternateAbilityId(nAbility)))
 				{
 					WriteChatColorf("[ %d: %s ]", USERCOLOR_WHO, pAbility->ID,
 						pCDBStr->GetString(pAbility->nName, eAltAbilityName));
@@ -4262,7 +4262,7 @@ void AltAbility(SPAWNINFO* pChar, char* szLine)
 
 			for (int nAbility = 0; nAbility < AA_CHAR_MAX_REAL; nAbility++)
 			{
-				if (ALTABILITY* pAbility = GetAAByIdWrapper(pLocalPC->GetAlternateAbilityId(nAbility)))
+				if (CAltAbilityData* pAbility = GetAAById(pLocalPC->GetAlternateAbilityId(nAbility)))
 				{
 					if ((pAltAdvManager->GetCalculatedTimer(pLocalPC, pAbility)) > 0)
 					{
@@ -4295,7 +4295,7 @@ void AltAbility(SPAWNINFO* pChar, char* szLine)
 	{
 		for (int nAbility = 0; nAbility < NUM_ALT_ABILITIES; nAbility++)
 		{
-			if (ALTABILITY* pAbility = GetAAByIdWrapper(nAbility))
+			if (CAltAbilityData* pAbility = GetAAById(nAbility))
 			{
 				const char* pName = pCDBStr->GetString(pAbility->nName, eAltAbilityName);
 				if (!_stricmp(pName, szName))
@@ -4368,7 +4368,7 @@ void AltAbility(SPAWNINFO* pChar, char* szLine)
 		// only search through the ones we have...
 		for (int nAbility = 0; nAbility < AA_CHAR_MAX_REAL; nAbility++)
 		{
-			if (ALTABILITY* pAbility = GetAAByIdWrapper(pLocalPC->GetAlternateAbilityId(nAbility), level))
+			if (CAltAbilityData* pAbility = GetAAById(pLocalPC->GetAlternateAbilityId(nAbility), level))
 			{
 				if (const char* pName = pCDBStr->GetString(pAbility->nName, eAltAbilityName))
 				{